# Optimize strutils.repeat