@Araq 2023-09-23
The benchmark is flawed anyway. I claim the most common usage is to produce a short string (2-8 chars) with spaces or tabs inside. A string of length 1000_000 is rare by the law of small numbers.